python/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
ultralytics==8.0.20
opencv-python==4.8.0.74
pyserial==3.5
torch==2.0.1
torchvision==0.15.2
numpy==1.24.0
Pillow==10.0.0
